Refactor runtime, get rid of gas syscalls and default syscalls
