+++ tests/conformance_tests/tests/conformance_runner.rs
@@ -58,10 +58,7 @@ lazy_static! {
         Regex::new(r"fil_1_storageminer-DeclareFaults-Ok-7").unwrap(),
         Regex::new(r"fil_1_storageminer-SubmitWindowedPoSt").unwrap(),
 
-        // Other faults (related to versioning, randomness, )
-        Regex::new(r"paych--update-ok--genesis").unwrap(),
-        Regex::new(r"paych--collect-ok--").unwrap(),
-        Regex::new(r"actor_exec--msg-apply-fail-actor-execution-illegal-arg--genesis").unwrap(),
+        // Other fault
         Regex::new(r"fail-insufficient-funds-for-transfer-in-inner-send--genesis").unwrap(),
     ];
 }
