+++ vm/interpreter/src/lib.rs
@@ -5,14 +5,12 @@
 extern crate lazy_static;
 
 mod default_runtime;
-mod default_syscalls;
 mod gas_block_store;
-mod gas_syscalls;
 mod gas_tracker;
 mod rand;
 mod vm;
+
 pub use self::default_runtime::*;
-pub use self::default_syscalls::DefaultSyscalls;
 pub use self::gas_tracker::*;
 pub use self::rand::*;
 pub use self::vm::*;
