+++ vm/interpreter/tests/transfer_test.rs
@@ -7,7 +7,7 @@ use blocks::TipsetKeys;
 use cid::multihash::{Blake2b256, Identity};
 use db::MemoryDB;
 use fil_types::{verifier::MockVerifier, NetworkVersion};
-use interpreter::{vm_send, ChainRand, DefaultRuntime, DefaultSyscalls};
+use interpreter::{vm_send, ChainRand, DefaultRuntime};
 use ipld_blockstore::BlockStore;
 use ipld_hamt::Hamt;
 use message::UnsignedMessage;
@@ -92,15 +92,12 @@ fn transfer_test() {
         .build()
         .unwrap();
 
-    let default_syscalls = DefaultSyscalls::<_, MockVerifier>::new(&store);
-
     let dummy_rand = ChainRand::new(TipsetKeys::new(vec![]));
     let registered = HashSet::new();
-    let mut runtime = DefaultRuntime::<_, _, _>::new(
+    let mut runtime = DefaultRuntime::<_, _, MockVerifier>::new(
         NetworkVersion::V0,
         &mut state,
         &store,
-        &default_syscalls,
         0,
         &message,
         0,
